reorganization of data prep code and first edits
