+++ tomcat_speech/data_prep/asist/asist_dataset_creation.py
@@ -68,10 +68,7 @@ class AsistDataset(Dataset):
             self.get_min_max_scales()
 
         self.skipped_files = []
-        # Tests gender classifier:print('start')
 
-        # self.x_acoustic, self.x_glove, self.x_speaker, self.x_utt_lengths = self.combine_acoustic_and_glove()
-        # Tests gender classifierprint('end')
         if transcript_type.lower() == "zoom":
             (
                 self.x_acoustic,
@@ -85,7 +82,6 @@ class AsistDataset(Dataset):
                 self.x_glove,
                 self.x_speaker,
                 self.x_utt_lengths,
-                # self.x_speaker_gender, #sa
             ) = self.combine_acoustic_and_glove_wd_level()
 
         # todo: we should get gender info on participants OR predict it
